Separate Common module by concerns
